+++ src/app/dashboard/commands-module/commands-module.component.ts
@@ -31,6 +31,7 @@ export class CommandsModuleComponent extends ModuleConfig implements OnInit {
     if (this.service.commands.length <= 0)
       await this.service.updateCommands();
     this.commands = this.service.commands;
+    this.commands.sort((a, b) => a.name > b.name ? 1 : -1);
 
     await super.init();
   }