+++ src/app/dashboard/plus-paywall/plus-paywall.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-plus-paywall',
+  templateUrl: './plus-paywall.component.html',
+  styleUrls: ['./plus-paywall.component.css']
+})
+export class PlusPaywallComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}