+++ src/assets/docs/changelog.md
@@ -7,6 +7,7 @@
 ---
 
 ### v0.2.1b
+`Upcoming`
 
 ![v0.2.1b Dashboard](assets/docs/img/dashboard-v0.2.1.png)
 
@@ -16,6 +17,7 @@
 
 **Change/Fix**
 - home -> smooth home page image
+- dashboard -> commands -> commands are now sorted alphabetically
 
 **Remove**